+++ tests/unit/components/select-2-customize-test.js
@@ -1,5 +1,3 @@
-/* globals andThen */
-
 /*
 	Test "customization" options like:
 	 * `optionLabelPath`