+++ tests/unit/components/select-2-enabled-option-test.js
@@ -1,5 +1,3 @@
-/* globals andThen */
-
 /*
   Test the `enabled=true/false` binding
  */